OCSADV-200: move VcsOtClient and VcsSyncDialog to jsky.app.ot.vcs
