+++ bundle/jsky.app.ot/src/main/scala/jsky/app/ot/osgi/Activator.scala
@@ -14,7 +14,7 @@ import edu.gemini.util.security.auth.ui.PasswordDialog
 import jsky.app.ot.OT
 import jsky.app.ot.gemini.obscat.CatalogQueryHistory
 import jsky.app.ot.plugin.{OtActionPlugin, OtViewerService}
-import jsky.app.ot.vcs2.VcsOtClient
+import jsky.app.ot.vcs.VcsOtClient
 import jsky.app.ot.viewer.ViewerService
 
 import org.osgi.framework.{ServiceRegistration, BundleContext, BundleActivator}
