+++ bundle/jsky.app.ot/src/main/scala/jsky/app/ot/viewer/open/ProgTableModel.scala
@@ -17,7 +17,7 @@ import edu.gemini.util.trpc.client.TrpcClient
 
 import jsky.app.ot.OT
 import jsky.app.ot.shared.spModel.util.DBProgramListFunctor
-import jsky.app.ot.vcs2.VcsOtClient
+import jsky.app.ot.vcs.VcsOtClient
 import jsky.app.ot.viewer.DBProgramChooserFilter
 import jsky.util.gui.DialogUtil
